Test fixes and ccno update
Made events have a static pointer to a ccno state field set in ClockManager instead of having it passed in as a const uint reference. Updated testing of event storage to decrement the event pointer after signaling to allow for getting data of the event just updated (since signal will automatically increment it).